requirements/dev.txt
Fabric==1.10.2
django-debug-toolbar==1.4
django-extensions==1.6.1